+++ test/govspeak_table_with_headers_test.rb
@@ -131,6 +131,39 @@ class GovspeakTableWithHeadersTest < Minitest::Test
 )
   end
 
+  def expected_outcome_for_table_headers_containing_links
+    %(
+<table>
+  <thead>
+    <tr>
+      <td></td>
+      <th scope="col">Second Column</th>
+      <th scope="col">Third Column</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th scope="row">Link contained in header <a rel="external" href="http://google.com">link1</a>
+</th>
+      <td>Cell</td>
+      <td>Cell</td>
+    </tr>
+    <tr>
+      <th scope="row">No link</th>
+      <td>Cell</td>
+      <td>Cell</td>
+    </tr>
+    <tr>
+      <th scope="row"><a rel="external" href="http://www.bbc.co.uk">Whole header is link</a>
+</th>
+      <td>Cell</td>
+      <td>Cell</td>
+    </tr>
+  </tbody>
+</table>
+)
+  end
+
   def document_body_with_hashes_for_all_headers
     @document_body_with_hashes_for_all_headers ||= Govspeak::Document.new(%(
 |                 |# Second Column  |# Third Column       |
@@ -176,6 +209,16 @@ class GovspeakTableWithHeadersTest < Minitest::Test
 ))
   end
 
+  def document_body_with_table_headers_containing_links
+    @document_body_with_table_headers_containing_links ||= Govspeak::Document.new(%(
+|                                                      | Second Column   | Third Column        |
+| ---------------------------------------------------- | --------------- | ------------------- |
+|# Link contained in header [link1](http://google.com) | Cell            | Cell                |
+|# No link                                             | Cell            | Cell                |
+|# [Whole header is link](http://www.bbc.co.uk)        | Cell            | Cell                |
+))
+  end
+
   test "Cells with |# are headers" do
     assert_equal document_body_with_hashes_for_all_headers.to_html, expected_outcome
   end
@@ -190,6 +233,10 @@ class GovspeakTableWithHeadersTest < Minitest::Test
 
   test "Table headers with a scope of row are only in the first column of the table" do
     assert_equal document_body_with_table_headers_in_the_wrong_place.to_html, expected_outcome_for_table_headers_in_the_wrong_place
+  end
+
+  test "Table headers with a scope of row can have embedded links" do
+    assert_equal document_body_with_table_headers_containing_links.to_html, expected_outcome_for_table_headers_containing_links
   end
 
   test "Table headers are not blank" do